requirements.txt
requests>=2.32.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.0,<1.0.0
langgraph>=0.2.0,<1.0.0
cachetools>=5.3.0,<6.0.0
tenacity>=8.2.0,<10.0.0
